Fix issue #4 by storing only 1 item per month in metafile-history
