+++ src/query.test.js
@@ -1,0 +1,8 @@
+import { queryMetafileHistory } from "./query.js"
+
+await testQueryCommitHistory();
+
+async function testQueryCommitHistory() {
+    const metafile = await queryMetafileHistory("com.valvesoftware.Steam")
+    console.log(metafile)
+}
